refactored factories out separately so faker module doesn't get included in prod bundle. it represented 28% of the bundle's code!
