+++ app/imports/api/league_pick_expected_wins/league_pick_expected_wins.js
@@ -1,7 +1,5 @@
 import { Mongo } from 'meteor/mongo';
 import { SimpleSchema } from 'meteor/aldeed:simple-schema';
-import { Factory } from 'meteor/dburles:factory';
-import log from '../../utils/log';
 
 export const LeaguePickExpectedWins = new Mongo.Collection('league_pick_expected_wins');
 
@@ -25,10 +23,4 @@ LeaguePickExpectedWins.deny({
   insert() { return true; },
   update() { return true; },
   remove() { return true; },
-});
-
-Factory.define('leaguePickExpectedWin', LeaguePickExpectedWins, {
-  leagueId: Factory.get('league'),
-}).after(leaguePickExpectedWin => {
-  log.debug('leaguePickExpectedWin factory created:', leaguePickExpectedWin);
 });
