+++ app/imports/api/league_teams/league_teams.js
@@ -1,10 +1,6 @@
 import { Mongo } from 'meteor/mongo';
 import { SimpleSchema } from 'meteor/aldeed:simple-schema';
-import { Factory } from 'meteor/dburles:factory';
-import faker from 'faker';
-import log from '../../utils/log';
 
-import '../leagues/leagues'; // needed for factory
 export const LeagueTeams = new Mongo.Collection('league_teams');
 
 LeagueTeams.schema = new SimpleSchema({
@@ -46,26 +42,4 @@ LeagueTeams.deny({
   insert() { return true; },
   update() { return true; },
   remove() { return true; },
-});
-
-Factory.define('leagueTeam', LeagueTeams, {
-  leagueId: Factory.get('league'),
-  cityName: () => faker.address.city(),
-  mascotName: () => faker.company.companyName(),
-  abbreviation: () => faker.lorem.words(1),
-  conference: 'NFC',
-  division: 'East',
-}).after(leagueTeam => {
-  log.debug('leagueTeam factory created:', leagueTeam);
-});
-
-Factory.define('awayLeagueTeam', LeagueTeams, {
-  leagueId: Factory.get('league'),
-  cityName: () => faker.address.city(),
-  mascotName: () => faker.company.companyName(),
-  abbreviation: () => faker.lorem.words(1),
-  conference: 'AFC',
-  division: 'South',
-}).after(awayLeagueTeam => {
-  log.debug('awayLeagueTeam factory created:', awayLeagueTeam);
 });
