+++ app/imports/api/pool_team_picks/pool_team_picks.js
@@ -1,11 +1,8 @@
 import { Mongo } from 'meteor/mongo';
 import { SimpleSchema } from 'meteor/aldeed:simple-schema';
 import { _ } from 'meteor/underscore';
-import { Factory } from 'meteor/dburles:factory';
-import log from '../../utils/log';
 
 import { LeagueTeams } from '../league_teams/league_teams';
-import { SeasonLeagueTeams } from '../season_league_teams/season_league_teams';
 import { PoolTeams } from '../pool_teams/pool_teams';
 import { Pools } from '../pools/pools';
 import { Seasons } from '../seasons/seasons';
@@ -196,12 +193,3 @@ if (Meteor.isServer) {
     },
   });
 }
-
-
-Factory.define('poolTeamPick', PoolTeamPicks, {
-  poolTeamId: Factory.get('poolTeam'),
-  leagueTeamId: Factory.get('leagueTeam'),
-  pickNumber() { return _.random(1, 32); },
-}).after(factory => {
-  log.debug('poolTeamPick factory created:', factory);
-});
