+++ app/imports/api/pool_team_picks/server/updater.tests.js
@@ -1,16 +1,17 @@
 /* eslint-env mocha */
 /* eslint-disable func-names, prefer-arrow-callback */
 
+import { assert } from 'chai';
 import { Factory } from 'meteor/dburles:factory';
 import log from '../../../utils/log';
-import '../../pool_teams/pool_teams'; // needed for factory
-import '../../league_teams/league_teams'; // needed for factory
-import '../../season_league_teams/season_league_teams'; // needed for factory
+
+import '../../pool_teams/pool_teams_factory';
+import '../../league_teams/league_teams_factory';
+import '../../season_league_teams/season_league_teams_factory';
+import '../../league_pick_expected_wins/league_pick_expected_wins_factory';
 
 import { PoolTeamPicks } from '../../pool_team_picks/pool_team_picks';
 import PoolTeamPicksUpdater from '../../pool_team_picks/server/updater';
-
-import { assert } from 'chai';
 
 describe('Pool Team Picks Updater', function () {
   this.timeout(5000);
