+++ app/imports/api/pools/pools.js
@@ -1,13 +1,6 @@
 import { Meteor } from 'meteor/meteor';
 import { Mongo } from 'meteor/mongo';
 import { SimpleSchema } from 'meteor/aldeed:simple-schema';
-import { Factory } from 'meteor/dburles:factory';
-import faker from 'faker';
-import log from '../../utils/log';
-
-import '../leagues/leagues'; // needed for factory
-import '../seasons/seasons'; // needed for factory
-import '../users/users'; // needed for factory
 import SeasonFinder from '../seasons/finder';
 
 export const Pools = new Mongo.Collection('pools');
@@ -95,12 +88,3 @@ if (Meteor.isServer) {
     },
   });
 }
-
-Factory.define('pool', Pools, {
-  leagueId: Factory.get('league'),
-  name: () => faker.lorem.words(),
-  commissionerUserId: Factory.get('user'),
-  latestSeasonId: Factory.get('season'),
-}).after(factory => {
-  log.debug('pool factory created:', factory);
-});
