+++ app/imports/api/pools/server/hooks.tests.js
@@ -1,14 +1,12 @@
 /* eslint-env mocha */
 /* eslint-disable func-names, prefer-arrow-callback */
 
+import { assert } from 'chai';
 import { Factory } from 'meteor/dburles:factory';
-import log from '../../../utils/log';
 
 import { PoolTeams } from '../../pool_teams/pool_teams';
 import { Pools } from '../pools';
 import './hooks';
-
-import { assert } from 'chai';
 
 describe('Pools Hooks', function () {
   this.timeout(5000);
