+++ app/imports/api/season_league_teams/server/updater.tests.js
@@ -1,16 +1,15 @@
 /* eslint-env mocha */
 /* eslint-disable func-names, prefer-arrow-callback */
 
+import { assert } from 'chai';
 import { Factory } from 'meteor/dburles:factory';
-import log from '../../../utils/log';
-import '../../seasons/seasons'; // needed for factory
-import '../../league_teams/league_teams'; // needed for factory
+
+import '../../seasons/seasons_factory';
+import '../../league_teams/league_teams_factory';
 
 import { SeasonLeagueTeams } from '../../season_league_teams/season_league_teams';
 import { Games } from '../../games/games';
 import SeasonLeagueTeamsUpdater from '../../season_league_teams/server/updater';
-
-import { assert } from 'chai';
 
 describe('Season League Teams Updater', () => {
   describe('Update Team Stats', () => {
