+++ app/imports/api/seasons/seasons.js
@@ -1,11 +1,8 @@
 import { Mongo } from 'meteor/mongo';
 import { SimpleSchema } from 'meteor/aldeed:simple-schema';
-import { _ } from 'meteor/underscore';
-import { Factory } from 'meteor/dburles:factory';
 import 'meteor/dburles:collection-helpers';
-import log from '../../utils/log';
 
-import { Leagues } from '../leagues/leagues'; // needed for factory
+import { Leagues } from '../leagues/leagues';
 
 export const Seasons = new Mongo.Collection('seasons');
 
@@ -35,16 +32,6 @@ Seasons.allow({
   insert() { return Meteor.isAppTest; },
   update() { return false; },
   remove() { return false; },
-});
-
-Factory.define('season', Seasons, {
-  leagueId: Factory.get('league'),
-  year: () => _.random(2000, 2014),
-  startDate: new Date(),
-  endDate: new Date(),
-  status: 'in progress',
-}).after(factory => {
-  log.debug('season factory created:', factory);
 });
 
 Seasons.helpers({
